# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Pessoas",
        "version": "0.1"
    },
    "paths": {
...
        "/api/v1/registrar": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Cadastro de usuário",
                "description": "registrar",
                "operationId": "registrar",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password",
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Leonardo"
                                    },
                                    "email": {
                                        "description": "Email",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "Password",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sucesso. Cadastro realizado com sucesso."
                    },
                    "400": {
                        "description": "Erro. Não foi possível realizar o cadastro."
                    }
                }
            }
        },
        "/api/v1/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login de usuário",
                "description": "Login",
                "operationId": "Login",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "Email",
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sucesso. Retorna o token de acesso."
                    },
                    "401": {
                        "description": "Não autorizado. Usuário ou senha inválidos."
                    }
                }
            }
        },
        "/api/v1/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logout de usuário",
                "description": "Logout",
                "operationId": "logout",
                "responses": {
                    "200": {
                        "description": "Sucesso. Logout realizado com sucesso."
                    },
                    "401": {
                        "description": "Não autorizado. Token de acesso inválido."
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/v1/user": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Retorna os dados do usuário logado",
                "description": "user",
                "operationId": "user",
                "responses": {
                    "200": {
                        "description": "Sucesso. Retorna os dados do usuário logado."
                    },
                    "401": {
                        "description": "Não autorizado. Token de acesso inválido."
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "passport": {
                "type": "apiKey",
                "description": "Laravel passport apiKey security.",
                "in": "header",
                "scheme": "http",
                "name": "bearer",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://localhost/oauth/authorize",
                        "tokenUrl": "http://localhost/oauth/token",
                        "refreshUrl": "http://localhost/token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    }
}